lift useMemoryGame up
